refactor(mediawiki): extract eval suite/result types from evals/runner.go (deslop #59)
